socket fixes, fix tank jumping

+++ TanksServer/Graphics/DrawBulletsStep.cs
@@ -8,6 +8,6 @@ internal sealed class DrawBulletsStep(BulletManager bullets) : IDrawStep
     public void Draw(PixelGrid buffer)
     {
         foreach (var position in bullets.GetAll().Select(b => b.Position.ToPixelPosition()))
-            buffer[position.X, position.Y] = true;
+            buffer[(ushort)position.X, (ushort)position.Y] = true;
     }
 }
